Move documentFile queries to separate file
