+++ lib/graphql/queries/package.js
@@ -1,5 +1,5 @@
 import gql from 'graphql-tag';
-import { IMAGE_DETAILS_FRAGMENT, LANGUAGE_DETAILS_FRAGMENT } from './common';
+import { DOCUMENT_DETAILS_FRAGMENT } from './document';
 
 /*----------------------------------------
  Fragments
@@ -46,65 +46,6 @@ export const PACKAGE_DETAILS_FRAGMENT = gql`
       }
     }
   }
-`;
-
-export const DOCUMENT_DETAILS_FRAGMENT = gql`
-  fragment documentDetails on DocumentFile {
-    id
-    language {
-      ...languageDetails
-    }
-    filetype
-    filename
-    filesize
-    status
-    content {
-      id
-      rawText
-      html
-      markdown
-    }
-    image {
-      ...imageDetails
-    }
-    url
-    # signedUrl
-    visibility
-    use {
-      id
-      name
-    }
-    tags {
-      id 
-      translations {
-        name
-        language {
-          ...languageDetails
-        }
-      }
-    }
-    bureaus {
-      id
-      name
-      abbr
-      offices {
-        id
-        name
-        abbr
-      }
-    }
-    categories {
-      id
-      translations {
-        name
-        language {
-          ...languageDetails
-        }
-      }
-    }
-  }
-  ${LANGUAGE_DETAILS_FRAGMENT}
-  ${IMAGE_DETAILS_FRAGMENT}
 `;
 
 /*----------------------------------------
@@ -172,17 +113,4 @@ export const PACKAGE_FILES_QUERY = gql`
     }
   }
   ${DOCUMENT_DETAILS_FRAGMENT}
-`;
-
-export const DOCUMENT_FILE_QUERY = gql`
-  query DocumentFile($id: ID!) {
-    documentFile(id: $id) {
-      id
-      filename
-      image {
-        ...imageDetails
-      }
-    }
-  }
-  ${IMAGE_DETAILS_FRAGMENT}
 `;
